+++ udiskie/udisks2.py
@@ -496,6 +496,21 @@ class Device(object):
                 return device
         return None
 
+    @property
+    def should_automount(self):
+        """Check if the device should be automounted."""
+        return bool(self._I.Block.property.HintAuto)
+
+    @property
+    def icon_name(self):
+        """Return the recommended device icon name."""
+        return self._I.Block.property.HintIconName or 'drive-removable-media'
+
+    @property
+    def symbolic_icon_name(self):
+        """Return the recommended device symbolic icon name."""
+        return self._I.Block.property.HintSymbolicIconName or 'drive-removable-media'
+
     # ----------------------------------------
     # Partition
     # ----------------------------------------