+++ CHANGES.rst
@@ -5,6 +5,7 @@ next
 ~~~~
 
 - use icon hints from udev settings in notifications
+- respect the automount flag for devices
 
 1.3.0
 ~~~~~