+++ CHANGES.rst
@@ -1,8 +1,8 @@
 CHANGELOG
 ---------
 
-next
-~~~~
+1.3.1
+~~~~~
 
 - use icon hints from udev settings in notifications
 - respect the automount flag for devices